requirements.txt
PyYAML
tqdm
colorama
